vllm_api/requirements.txt
fastapi
uvicorn
sentry-sdk
pydantic-settings
vllm
